+++ client/src/assets/pages/Stripe.jsx
@@ -1,7 +1,23 @@
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
 import React from "react";
 import { loadStripe } from "@stripe/stripe-js";
 import { Elements } from "@stripe/react-stripe-js";
-import CheckoutForm from "../components/Stripe/index.jsx";
+import CheckoutForm from "../components/Stripe/CheckoutForm";
 import "../components/Stripe/stripe.css";
 import CCInfoInput from "../components/Stripe/CCinput.jsx"
